MonitorRect now signed type
git-svn-id: https://svn.vis.uni-stuttgart.de/utilities/vislib/trunk@1031 3ec011a9-ccf5-4689-8f2f-5f4aa021d377

+++ sys/include/vislib/SystemInformation.h
@@ -41,7 +41,7 @@ namespace sys {
         };
 
         /** This dimension defines a monitor size. */
-        typedef vislib::math::Rectangle<DWORD> MonitorRect;
+        typedef vislib::math::Rectangle<long> MonitorRect;
 
         /** Array of MonitorRects. */
         typedef vislib::Array<MonitorRect> MonitorRectArray;
